update semantics of Pageable List components
